+++ server-common/src/test/java/ee/cyber/cdoc2/server/TestingConfiguration.java
@@ -15,11 +15,10 @@ import org.apache.hc.client5.http.ssl.SSLConnectionSocketFactoryBuilder;
 import org.apache.hc.client5.http.ssl.TrustAllStrategy;
 import org.apache.hc.client5.http.impl.classic.HttpClients;
 import org.apache.hc.core5.ssl.SSLContextBuilder;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.web.client.RestClient;
 
 
 /**
@@ -29,12 +28,11 @@ import org.springframework.web.client.RestTemplate;
 public class TestingConfiguration {
 
     /**
-     * @param builder the REST template builder
      * @return a REST template that trusts all hosts it connects to
      */
     @Bean(name = "trustAllNoClientAuth")
     @SneakyThrows
-    public RestTemplate getTrustAllRestTemplate(RestTemplateBuilder builder) {
+    public RestClient getTrustAllRestTemplate() {
         try {
             SSLContext sslContext = SSLContextBuilder
                 .create()
@@ -51,8 +49,8 @@ public class TestingConfiguration {
                 .setConnectionManager(cm)
                 .build();
 
-            return builder
-                .requestFactory(() -> new HttpComponentsClientHttpRequestFactory(client))
+            return RestClient.builder()
+                .requestFactory(new HttpComponentsClientHttpRequestFactory(client))
                 .build();
         } catch (NoSuchAlgorithmException | KeyStoreException | KeyManagementException e) {
             throw new GeneralSecurityException("Failed to create SSL context: " + e.getMessage());
@@ -60,12 +58,11 @@ public class TestingConfiguration {
     }
 
     /**
-     * @param builder the REST template builder
      * @return a REST template with client authentication and trusts all hosts strategy
      */
     @Bean(name = "trustAllWithClientAuth")
     @SneakyThrows
-    public RestTemplate getTrustAllWithClientAuthRestTemplate(RestTemplateBuilder builder) {
+    public RestClient getTrustAllWithClientAuthRestTemplate() {
         try {
             SSLContext sslContext = SSLContextBuilder
                 .create()
@@ -87,8 +84,8 @@ public class TestingConfiguration {
                 .setConnectionManager(cm)
                 .build();
 
-            return builder
-                .requestFactory(() -> new HttpComponentsClientHttpRequestFactory(client))
+            return RestClient.builder()
+                .requestFactory(new HttpComponentsClientHttpRequestFactory(client))
                 .build();
         } catch (NoSuchAlgorithmException | KeyStoreException | KeyManagementException e) {
             throw new GeneralSecurityException("Failed to create SSL context: " + e.getMessage());